refactor testcontainers to separate masternode
