+++ .github/workflows/ci.yml
@@ -20,5 +20,5 @@ jobs:
       - run: npm run build
       - run: npm run standard
 
-      - run: npx --no-install jest --ci --coverage --forceExit
+      - run: npm run test:ci
       - run: npx codecov
