+++ .idea/jellyfish.iml
@@ -6,6 +6,7 @@
       <excludeFolder url="file://$MODULE_DIR$/.parcel-cache" />
       <excludeFolder url="file://$MODULE_DIR$/website/.docusaurus" />
       <excludeFolder url="file://$MODULE_DIR$/website/build" />
+      <excludeFolder url="file://$MODULE_DIR$/coverage" />
       <excludePattern pattern="*dist*" />
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
