+++ jest.config.js
@@ -2,9 +2,5 @@ module.exports = {
   projects: [
     '<rootDir>/packages/*'
   ],
-  testTimeout: 240000,
-  collectCoverageFrom: [
-    'packages/**/*.{ts,js}',
-    '!packages/**/__tests__/**'
-  ]
+  testTimeout: 240000
 }
