+++ packages/testcontainers/src/index.ts
@@ -1,12 +1,9 @@
 export { DockerOptions } from 'dockerode'
 
-export { StartOptions, DeFiDContainer, DeFiDRpcError } from './chains/container'
-export { MainNetContainer } from './chains/main_net_container'
-export { TestNetContainer } from './chains/test_net_container'
-export {
-  RegTestContainer, MasterNodeRegTestContainer
-} from './chains/reg_test_container'
+export * from './testkeys'
 
-export {
-  KeyPair, MasterNodeKey, GenesisKeys
-} from './testkeys'
+export * from './chains/container'
+export * from './chains/main_net_container'
+export * from './chains/test_net_container'
+export * from './chains/reg_test_container/index'
+export * from './chains/reg_test_container/masternode'
